requirements.txt
PyAudio==0.2.7
numpy==1.13.3
scikit-learn==0.19.0
scikits.talkbox==0.2.5
scipy==0.19.1
